Remove unused migrators
